feat(profile): redesign profile page layout
