+++ components/UI/index.ts
@@ -1,3 +1,4 @@
 export { default as InputField } from './InputField';
 export { default as GenericInput } from './GenericInput';
 export { default as GenericModal } from '../Modals/GenericModal';
+export { default as StatusLabel } from './StatusLabel';
